Revert "refactor(labware-library): Add mobile submenus (#3401)"
This reverts commit a3224e008d2b29f524f54b21ed329f922a04ae8a.
