+++ labware-library/src/components/website-navigation/MainNav.js
@@ -2,15 +2,20 @@
 import * as React from 'react'
 import Logo from './Logo'
 import { NavList } from './NavList'
-import MobileNav from './MobileNav'
+import MenuButton from './MenuButton'
 import styles from './styles.css'
 
-export function MainNav() {
+import type { MobileNavProps } from './types'
+
+export function MainNav(props: MobileNavProps) {
   return (
     <div className={styles.main_nav_contents}>
       <Logo />
       <NavList />
-      <MobileNav />
+      <MenuButton
+        onMobileClick={props.onMobileClick}
+        isMobileOpen={props.isMobileOpen}
+      />
     </div>
   )
 }
