--- labware-library/src/components/website-navigation/MobileContent.js
@@ -1,33 +1,0 @@
-// @flow
-import * as React from 'react'
-import NavLink from './NavLink'
-import styles from './styles.css'
-
-import type { Submenu } from './types'
-
-type Props = {|
-  ...Submenu,
-|}
-
-export default function MobileContent(props: Props) {
-  const { links } = props
-  return (
-    <ul className={styles.mobile_content}>
-      {links.map((link, index) => (
-        <li key={index}>
-          <NavLink url={link.url} name={link.name} />
-        </li>
-      ))}
-
-      {props.bottomLink && (
-        <li>
-          <NavLink
-            url={props.bottomLink.url}
-            name={props.bottomLink.name}
-            cta
-          />
-        </li>
-      )}
-    </ul>
-  )
-}
