+++ labware-library/src/components/website-navigation/NavLink.js
@@ -5,14 +5,9 @@ import styles from './styles.css'
 
 import type { Link } from './types'
 
-type Props = {|
-  ...Link,
-  className?: string,
-|}
-
-export default function NavLink(props: Props) {
+export default function NavLink(props: Link) {
   return (
-    <div className={cx(styles.link_group, props.className)}>
+    <div className={styles.link_group}>
       <a
         href={props.url}
         className={cx(styles.link_title, { [styles.link_cta]: props.cta })}
@@ -20,6 +15,7 @@ export default function NavLink(props: Props) {
         rel="noopener noreferrer"
       >
         {props.name}
+        {props.cta && <span>&nbsp; &gt;</span>}
       </a>
       {props.description && (
         <div className={styles.link_description}>{props.description}</div>
