+++ labware-library/src/components/website-navigation/index.js
@@ -5,3 +5,4 @@
 export * from './SubdomainNav'
 export { MainNav } from './MainNav'
 export { NavList } from './NavList'
+export { MobileNav } from './MobileNav'
