+++ labware-library/src/components/website-navigation/nav-data.js
@@ -93,7 +93,6 @@ export const protocolLinkProps: ProtocolLinks = {
   bottomLink: {
     name: 'Request a free custom protocol',
     url: 'https://opentrons.com/request-protocol',
-    cta: true,
   },
 }
 
@@ -127,7 +126,6 @@ export const supportLinkProps: SupportLinks = {
   support: {
     name: 'Contact Support',
     url: 'http://opentrons.com/contact',
-    cta: true,
   },
 }
 
@@ -139,11 +137,9 @@ export const salesLinkProps: SalesLinks = {
   sales: {
     name: 'Contact Sales',
     url: 'http://opentrons.com/contact',
-    cta: true,
   },
   demo: {
     name: 'Schedule Demo',
     url: 'http://opentrons.com/demo',
-    cta: true,
   },
 }
