+++ labware-library/src/components/website-navigation/styles.css
@@ -9,7 +9,7 @@
   --spacing-subdomain: 2.5rem;
   --spacing-nav: 1.75rem;
   --spacing-submenu: 1.75rem;
-  --spacing-link: 1.75rem;
+  --spacing-link: 1.25rem;
 
   --dropdown-base: {
     position: absolute;
@@ -55,7 +55,6 @@
   --bd-button: var(--bd-width-default) solid var(--c-blue);
 }
 
-/* Subdomain Nav (wrapper and container responsive styles in Nav/) */
 .subdomain_nav_contents {
   @apply --flex-end;
 
@@ -71,7 +70,6 @@
   cursor: pointer;
 }
 
-/* Main Nav (wrapper and container responsive styles in Nav/) */
 .main_nav_contents {
   @apply --flex-between;
 
@@ -81,7 +79,7 @@
 }
 
 .logo {
-  height: 2rem;
+  height: 100%;
 }
 
 .nav_list {
@@ -109,7 +107,15 @@
   color: var(--c-black);
 }
 
-/* Dropdown Containers */
+.clickable_icon {
+  cursor: pointer;
+
+  &:active,
+  &:focus {
+    outline: none;
+  }
+}
+
 .dropdown_small {
   @apply --font-body-2-dark;
   @apply --dropdown-base;
@@ -171,7 +177,6 @@
   }
 }
 
-/* Dropdown Content */
 .dropdown_content {
   display: flex;
   padding: var(--spacing-submenu);
@@ -182,20 +187,19 @@
   flex: 1 1 var(--size-50p);
 }
 
-.submenu_title {
-  @apply --font-body-2-dark;
-
+.link_group {
   line-height: 1.25rem;
-  flex-basis: 8rem;
-  padding: 0 var(--spacing-5);
-  text-transform: uppercase;
+  margin: 0 var(--spacing-link) var(--spacing-link) 0;
 }
 
-/*  Support & Sales Submenu Desktop */
 .support_menu {
   padding: var(--spacing-submenu);
   padding-bottom: var(--spacing-1);
   flex: 1 1 var(--size-two-thirds);
+}
+
+.support_content {
+  display: flex;
 }
 
 .sales_menu {
@@ -204,47 +208,24 @@
   flex: 1 1 var(--size-third);
 }
 
-.support_content {
-  display: flex;
-}
+.submenu_title {
+  @apply --font-body-2-dark;
 
-/* NavLink */
-.link_group {
-  line-height: 1.25rem;
-  text-align: left;
-  padding-bottom: var(--spacing-link);
-  padding-left: 4rem;
-
-  &.support_nav_link {
-    padding-bottom: var(--spacing-5);
-    text-align: left;
-  }
+  flex: 0 0 100%;
+  margin-bottom: var(--spacing-5);
+  text-transform: uppercase;
 }
 
 .link_title {
   @apply --link-text;
 
-  text-align: left;
   color: var(--c-black);
 
   &:hover {
     color: var(--c-blue);
   }
-
-  /* Only add caret to base NavLinks on Mobile */
-  &::after {
-    content: '\00a0 \3E';
-  }
 }
 
-.link_description {
-  font-size: var(--fs-body-1);
-  font-weight: var(--fw-regular);
-  color: var(--c-nav-gray);
-  text-transform: none;
-}
-
-/*  Nav Link CTA */
 .link_cta {
   color: var(--c-blue);
 }
@@ -263,6 +244,11 @@
   }
 }
 
+.link_description {
+  font-size: var(--fs-body-1);
+  text-transform: none;
+}
+
 .bottom_link {
   @apply --flex-start;
   @apply --bottom-link;
@@ -275,20 +261,10 @@
   @apply --center-children;
 }
 
-/* Mobile Button / Toggle */
-.clickable_icon {
-  cursor: pointer;
-
-  &:active,
-  &:focus {
-    outline: none;
-  }
-}
-
 .menu_button {
   @apply --transition-color;
 
-  flex: 0 0 3rem;
+  flex: 0 0 4rem;
   color: var(--c-font-dark);
 
   &:hover {
@@ -296,24 +272,23 @@
   }
 }
 
-/* Mobile List */
 .mobile_nav {
   position: fixed;
-  top: var(--size-mobile-nav);
-  left: 0;
-  right: 0;
   width: 100%;
   padding: 0;
+  padding-bottom: var(--spacing-5);
   background-color: var(--c-white);
   z-index: 900;
-  border-bottom: var(--bd-light);
+
+  /* Added this for contrast with page - does not match design */
+  box-shadow: var(--shadow-1);
 }
 
 .mobile_nav_item {
   @apply --flex-start;
 
   padding: var(--spacing-7);
-  height: 4.5rem;
+  height: var(--size-2);
   font-weight: var(--fw-semibold);
   cursor: pointer;
 
@@ -322,14 +297,13 @@
   }
 }
 
-/* Mobile Menu container */
 .mobile_menu {
   position: fixed;
-  top: var(--size-mobile-nav);
+  bottom: 0;
   right: 0;
   width: 100%;
-  min-height: calc(100vh - var(--size-mobile-nav));
-  margin: 0 -105% 0 0;
+  min-height: calc(100vh - var(--size-main-nav));
+  margin: 0 -100% 0 0;
   background-color: var(--c-white);
   box-shadow: var(--shadow-1);
   z-index: 950;
@@ -340,11 +314,10 @@
   }
 }
 
-/* Top Mobile Menu bar */
 .mobile_menu_heading {
   @apply --flex-start;
 
-  height: 4.5rem;
+  height: 5rem;
   border-bottom: var(--bd-light);
 }
 
@@ -354,86 +327,20 @@
   width: var(--size-2);
 }
 
-/* Top Mobile Menu text */
 .mobile_menu_title {
   width: 100%;
-  padding-left: 5rem;
-  text-align: left;
+  text-align: center;
   font-size: var(--fs-default);
   font-weight: var(--fw-semibold);
 }
 
-.mobile_content {
-  padding: var(--spacing-5);
-  padding-bottom: 0;
-}
-
-/*  Support & Sales Submenu Mobile */
-
-.support_mobile_content {
-  display: flex;
-  flex-direction: column;
-}
-
-.sales_group {
-  display: flex;
-  flex-basis: 100%;
-  padding-top: var(--spacing-6);
-  border-bottom: var(--bd-light);
-}
-
-.support_group {
-  display: flex;
-  flex-basis: 100%;
-  padding-top: var(--spacing-6);
-}
-
-.support_submenu {
-  display: flex;
-}
-
 @media (--medium) {
-  .logo {
-    height: 2.5rem;
-  }
-
-  .link_group {
-    text-align: center;
-    padding-left: 0;
-  }
-
-  .mobile_nav {
-    top: var(--size-main-nav);
-  }
-
   .mobile_nav_item {
     justify-content: center;
   }
 
-  .mobile_menu_title {
-    padding: 0;
-    text-align: center;
-  }
-
   .mobile_menu {
-    top: var(--size-main-nav);
-    min-height: 29.5rem;
-  }
-
-  .support_mobile_content {
-    flex-direction: row;
-  }
-
-  .support_group {
-    flex-basis: 60%;
-    display: flex;
-  }
-
-  .sales_group {
-    flex-basis: 40%;
-    display: flex;
-    padding-bottom: 0;
-    border-bottom: none;
+    min-height: 30rem;
   }
 }
 
@@ -445,35 +352,5 @@
 
   .nav_list {
     @apply --flex-end;
-  }
-
-  .active {
-    color: var(--c-black);
-  }
-
-  .link_group {
-    line-height: 1.25rem;
-    padding: 0 var(--spacing-link) var(--spacing-link) 0;
-    text-align: left;
-  }
-
-  .link_title {
-    display: block;
-    padding: 0;
-    text-align: left;
-
-    &::after {
-      content: '';
-    }
-  }
-
-  .link_cta::after {
-    content: '\00a0 \3E';
-  }
-
-  .submenu_title {
-    flex: 0 0 100%;
-    padding: 0;
-    margin-bottom: var(--spacing-5);
   }
 }
